ref(metrics): Split utils to avoid chartcuterie build failing
