+++ static/app/utils/metrics/index.tsx
@@ -1,8 +1,6 @@
-import {useEffect, useMemo, useState} from 'react';
 import {InjectedRouter} from 'react-router';
 import moment from 'moment';
 
-import {ApiResult} from 'sentry/api';
 import {
   DateTimeObject,
   getDiffInMinutes,
@@ -12,29 +10,13 @@ import {t} from 'sentry/locale';
 import {
   MetricsApiRequestMetric,
   MetricsApiRequestQuery,
-  MetricsApiResponse,
   MetricsGroup,
 } from 'sentry/types/metrics';
 import {defined, formatBytesBase2, formatBytesBase10} from 'sentry/utils';
 import {parseFunction} from 'sentry/utils/discover/fields';
 import {formatPercentage, getDuration} from 'sentry/utils/formatters';
-import {ApiQueryKey, useApiQuery} from 'sentry/utils/queryClient';
-import useOrganization from 'sentry/utils/useOrganization';
 
-import {DateString, PageFilters} from '../types/core';
-
-// TODO(ddm): reuse from types/metrics.tsx
-type MetricMeta = {
-  mri: string;
-  name: string;
-  operations: string[];
-  type: string;
-  unit: string;
-};
-
-interface Options {
-  useCases?: UseCase[];
-}
+import {PageFilters} from '../../types/core';
 
 export enum MetricDisplayType {
   LINE = 'line',
@@ -45,95 +27,11 @@ export enum MetricDisplayType {
 
 export const defaultMetricDisplayType = MetricDisplayType.LINE;
 
-const DEFAULT_USE_CASES = ['sessions', 'transactions', 'custom'];
+export type UseCase = 'sessions' | 'transactions' | 'custom';
 
-export function useMetricsMeta(
-  projects: PageFilters['projects'],
-  options?: Options
-): {data: Record<string, MetricMeta>; isLoading: boolean} {
-  const {slug} = useOrganization();
-  const enabledUseCases = options?.useCases ?? DEFAULT_USE_CASES;
-
-  const getKey = (useCase: UseCase): ApiQueryKey => {
-    return [
-      `/organizations/${slug}/metrics/meta/`,
-      {query: {useCase, project: projects}},
-    ];
-  };
-
-  const commonOptions = {
-    staleTime: Infinity,
-  };
-
-  const sessionsMeta = useApiQuery<MetricMeta[]>(getKey('sessions'), {
-    ...commonOptions,
-    enabled: enabledUseCases.includes('sessions'),
-  });
-  const txnsMeta = useApiQuery<MetricMeta[]>(getKey('transactions'), {
-    ...commonOptions,
-    enabled: enabledUseCases.includes('transactions'),
-  });
-  const customMeta = useApiQuery<MetricMeta[]>(getKey('custom'), {
-    ...commonOptions,
-    enabled: enabledUseCases.includes('custom'),
-  });
-
-  const combinedMeta = useMemo<Record<string, MetricMeta>>(() => {
-    return [
-      ...(sessionsMeta.data ?? []),
-      ...(txnsMeta.data ?? []),
-      ...(customMeta.data ?? []),
-    ].reduce((acc, metricMeta) => {
-      return {...acc, [metricMeta.mri]: metricMeta};
-    }, {});
-  }, [sessionsMeta.data, txnsMeta.data, customMeta.data]);
-
-  return {
-    data: combinedMeta,
-    isLoading: sessionsMeta.isLoading || txnsMeta.isLoading || customMeta.isLoading,
-  };
-}
-
-type MetricTag = {
+export type MetricTag = {
   key: string;
 };
-
-export function useMetricsTags(
-  mri: string | undefined,
-  projects: PageFilters['projects']
-) {
-  const {slug} = useOrganization();
-  const useCase = getUseCaseFromMRI(mri || '');
-  return useApiQuery<MetricTag[]>(
-    [
-      `/organizations/${slug}/metrics/tags/`,
-      {query: {metric: mri, useCase, project: projects}},
-    ],
-    {
-      enabled: !!mri,
-      staleTime: Infinity,
-    }
-  );
-}
-
-export function useMetricsTagValues(
-  mri: string,
-  tag: string,
-  projects: PageFilters['projects']
-) {
-  const {slug} = useOrganization();
-  const useCase = getUseCaseFromMRI(mri);
-  return useApiQuery<MetricTag[]>(
-    [
-      `/organizations/${slug}/metrics/tags/${tag}/`,
-      {query: {metric: mri, useCase, project: projects}},
-    ],
-    {
-      staleTime: Infinity,
-      enabled: !!tag,
-    }
-  );
-}
 
 export type MetricsQuery = {
   datetime: PageFilters['datetime'];
@@ -144,45 +42,6 @@ export type MetricsQuery = {
   op?: string;
   query?: string;
 };
-
-export function useMetricsData({
-  mri,
-  op,
-  datetime,
-  projects,
-  environments,
-  query,
-  groupBy,
-}: MetricsQuery) {
-  const organization = useOrganization();
-
-  const useNewMetricsLayer = organization.features.includes(
-    'metrics-api-new-metrics-layer'
-  );
-
-  const field = op ? `${op}(${mri})` : mri;
-
-  const queryToSend = getMetricsApiRequestQuery(
-    {
-      field,
-      query: `${query}`,
-      groupBy,
-    },
-    {datetime, projects, environments},
-    {useNewMetricsLayer}
-  );
-
-  return useApiQuery<MetricsApiResponse>(
-    [`/organizations/${organization.slug}/metrics/data/`, {query: queryToSend}],
-    {
-      retry: 0,
-      staleTime: 0,
-      refetchOnReconnect: true,
-      refetchOnWindowFocus: true,
-      refetchInterval: data => getRefetchInterval(data, queryToSend.interval),
-    }
-  );
-}
 
 export function getMetricsApiRequestQuery(
   {field, query, groupBy}: MetricsApiRequestMetric,
@@ -209,73 +68,6 @@ export function getMetricsApiRequestQuery(
   return {...queryToSend, ...overrides};
 }
 
-function getRefetchInterval(
-  data: ApiResult | undefined,
-  interval: string
-): number | false {
-  // no data means request failed - don't refetch
-  if (!data) {
-    return false;
-  }
-  if (interval === '10s') {
-    // refetch every 10 seconds
-    return 10 * 1000;
-  }
-  // refetch every 60 seconds
-  return 60 * 1000;
-}
-
-// Wraps useMetricsData and provides two additional features:
-// 1. return data is undefined only during the initial load
-// 2. provides a callback to trim the data to a specific time range when chart zoom is used
-export function useMetricsDataZoom(props: MetricsQuery) {
-  const [metricsData, setMetricsData] = useState<MetricsApiResponse | undefined>();
-  const {data: rawData, isLoading, isError, error} = useMetricsData(props);
-
-  useEffect(() => {
-    if (rawData) {
-      setMetricsData(rawData);
-    }
-  }, [rawData]);
-
-  const trimData = (start, end): MetricsApiResponse | undefined => {
-    if (!metricsData) {
-      return metricsData;
-    }
-    // find the index of the first interval that is greater than the start time
-    const startIndex = metricsData.intervals.findIndex(interval => interval >= start) - 1;
-    const endIndex = metricsData.intervals.findIndex(interval => interval >= end);
-
-    if (startIndex === -1 || endIndex === -1) {
-      return metricsData;
-    }
-
-    return {
-      ...metricsData,
-      intervals: metricsData.intervals.slice(startIndex, endIndex),
-      groups: metricsData.groups.map(group => ({
-        ...group,
-        series: Object.fromEntries(
-          Object.entries(group.series).map(([seriesName, series]) => [
-            seriesName,
-            series.slice(startIndex, endIndex),
-          ])
-        ),
-      })),
-    };
-  };
-
-  return {
-    data: metricsData,
-    isLoading,
-    isError,
-    error,
-    onZoom: (start: DateString, end: DateString) => {
-      setMetricsData(trimData(start, end));
-    },
-  };
-}
-
 // Wraps getInterval since other users of this function, and other metric use cases do not have support for 10s granularity
 export function getMetricsInterval(dateTimeObj: DateTimeObject, useCase: UseCase) {
   const interval = getInterval(dateTimeObj, 'metrics');
@@ -298,8 +90,6 @@ export function getDateTimeParams({start, end, period}: PageFilters['datetime'])
     ? {statsPeriod: period}
     : {start: moment(start).toISOString(), end: moment(end).toISOString()};
 }
-
-type UseCase = 'sessions' | 'transactions' | 'custom';
 
 const metricTypeToReadable = {
   c: t('counter'),
