+++ static/app/utils/metrics/useMetricsTagValues.tsx
@@ -1,0 +1,23 @@
+import {PageFilters} from 'sentry/types';
+import {getUseCaseFromMRI, MetricTag} from 'sentry/utils/metrics';
+import {useApiQuery} from 'sentry/utils/queryClient';
+import useOrganization from 'sentry/utils/useOrganization';
+
+export function useMetricsTagValues(
+  mri: string,
+  tag: string,
+  projects: PageFilters['projects']
+) {
+  const {slug} = useOrganization();
+  const useCase = getUseCaseFromMRI(mri);
+  return useApiQuery<MetricTag[]>(
+    [
+      `/organizations/${slug}/metrics/tags/${tag}/`,
+      {query: {metric: mri, useCase, project: projects}},
+    ],
+    {
+      staleTime: Infinity,
+      enabled: !!tag,
+    }
+  );
+}
